Move Qwen linear trace out of main
