+++ crates/runner/src/main.rs
@@ -68,7 +68,6 @@ use policy::{
     validate_specprefill_flags,
 };
 use profiling::MetalProfileScope;
-use qwen35::state::ModelState;
 use qwen35_kv_trace::trace_kv_cache;
 use qwen35_runtime::{
     check_qwen35_vram, load_qwen35_engine, load_qwen35_startup, qwen35_oracle_script_path,
@@ -80,13 +79,11 @@ use qwen35_runtime::{
 use qwen35_trace::{
     trace_component_input_layer, trace_component_layer, trace_component_linear_layer,
     trace_component_linear_state_layer, trace_persistent_input_layer,
-    trace_persistent_full_attn_layer, trace_persistent_linear_state_layer,
+    trace_persistent_full_attn_layer, trace_persistent_linear_layer,
+    trace_persistent_linear_state_layer,
 };
-use qwen35_trace_utils::{bf16_residual_sum, bf16_round, sigmoid_fast};
 use registry::{Backend, FamilyParams, GpuArch, ModelFamily, ModelVariant};
 use supersonic_core::backend::{BackendChoice, BACKEND_CHOICES};
-use tensor_bytes::{bf16_bytes_to_f32 as decode_bf16_le, f32_bytes_to_f32 as decode_f32_le};
-
 #[derive(Parser)]
 #[command(name = "supersonic", about = "SuperSonic — optimized LLM inference")]
 pub(crate) struct Cli {
@@ -1640,699 +1637,5 @@ fn main() -> Result<()> {
         }
     }
 
-    Ok(())
-}
-
-fn trace_persistent_linear_layer(
-    engine: &mut DecodeEngine,
-    trace_layer: usize,
-    token_ids: &[u32],
-    trace_tokens: &[u32],
-    seqlen_offset: usize,
-    ordinal: usize,
-    kv_chunk_size: usize,
-    prefill_chunk_size: usize,
-    use_4b_kernel: bool,
-) -> Result<()> {
-    let text_config = engine.weights().config.clone();
-    anyhow::ensure!(
-        !text_config.is_full_attention(trace_layer),
-        "layer {trace_layer} is not a linear-attention layer"
-    );
-
-    let native_hidden = engine.decode_step_batch_trace_hidden_after_layers(
-        trace_tokens,
-        seqlen_offset,
-        trace_layer,
-        0,
-    )?;
-
-    let mut replay_state = ModelState::new(&text_config, ordinal)
-        .map_err(|e| anyhow::anyhow!("persistent linear replay state init: {e}"))?;
-    let replay = prefill_engine::prefill(
-        engine.weights(),
-        &mut replay_state,
-        engine.rotary(),
-        token_ids,
-        ordinal,
-        kv_chunk_size,
-        prefill_chunk_size,
-        false,
-        use_4b_kernel,
-        true,
-        None,
-    )?;
-    let replay_hidden = if trace_layer == 0 {
-        None
-    } else {
-        Some(
-            replay
-                .layer_hidden_trace
-                .as_ref()
-                .and_then(|layers| layers.get(trace_layer - 1))
-                .ok_or_else(|| {
-                    anyhow::anyhow!("missing replay hidden trace for layer {trace_layer}")
-                })?,
-        )
-    };
-    let replay_layer = replay_state
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing replay layer {trace_layer}"))?;
-    let replay_conv = replay_layer
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing replay conv state for layer {trace_layer}"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("replay conv D2H layer {trace_layer}: {e}"))?;
-    let replay_recurrent = replay_layer
-        .recurrent_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing replay recurrent state for layer {trace_layer}"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("replay recurrent D2H layer {trace_layer}: {e}"))?;
-    let replay_hidden_out = replay
-        .layer_hidden_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| {
-            anyhow::anyhow!("missing replay output hidden trace for layer {trace_layer}")
-        })?;
-    let replay_attn = replay
-        .layer_attn_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay attn trace for layer {trace_layer}"))?;
-    let replay_post = replay
-        .layer_post_attn_norm_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay post-attn trace for layer {trace_layer}"))?;
-    let replay_swiglu = replay
-        .layer_mlp_swiglu_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay swiglu trace for layer {trace_layer}"))?;
-    let replay_mlp_out = replay
-        .layer_mlp_out_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay mlp-out trace for layer {trace_layer}"))?;
-
-    let prefix_ids = token_ids
-        .get(..token_ids.len().saturating_sub(1))
-        .ok_or_else(|| anyhow::anyhow!("missing prefix token ids for persistent linear trace"))?;
-    engine.rebuild_prefill_state(prefix_ids, true)?;
-    let pre_step_conv = engine
-        .state_for_batch(0)
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing pre-step layer {trace_layer}"))?
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing pre-step conv state for layer {trace_layer}"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("pre-step conv D2H layer {trace_layer}: {e}"))?;
-    engine.set_hidden_from_bytes(&native_hidden)?;
-    let (native_comp_trace, native_comp_conv, native_comp_recurrent, native_comp_hidden) =
-        engine.component_trace_linear_layer_from_current_hidden(trace_layer)?;
-    engine.rebuild_prefill_state(prefix_ids, true)?;
-    engine.set_hidden_from_bytes(&native_hidden)?;
-    let native_comp_layer = engine
-        .component_trace_full_layer_from_current_hidden_with_seqlen(trace_layer, seqlen_offset)?;
-
-    engine.rebuild_prefill_state(prefix_ids, true)?;
-    let native_hidden_out = engine.decode_step_batch_trace_hidden_after_layers(
-        trace_tokens,
-        seqlen_offset,
-        trace_layer + 1,
-        0,
-    )?;
-    let native_partial_conv = engine
-        .state_for_batch(0)
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing native partial layer {trace_layer}"))?
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| {
-            anyhow::anyhow!("missing native partial conv state for layer {trace_layer}")
-        })?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("native partial conv D2H layer {trace_layer}: {e}"))?;
-    let cfg = engine.weights().config.clone();
-    let qkv_dim = cfg.linear_num_key_heads * cfg.linear_key_head_dim * 2
-        + cfg.linear_num_value_heads * cfg.linear_value_head_dim;
-    let z_dim = cfg.linear_num_value_heads * cfg.linear_value_head_dim;
-    let val_dim = cfg.linear_num_value_heads * cfg.linear_value_head_dim;
-    let nv = cfg.linear_num_value_heads;
-    let intermediate = cfg.intermediate_size;
-    let (native_qkv_proj, native_z_proj, native_b_proj, native_a_proj) =
-        engine.trace_persistent_linear_proj_buf_after_layers(0, qkv_dim, z_dim, nv)?;
-    let native_gated = engine.trace_persistent_linear_gated_after_layers(0, val_dim)?;
-    let (native_post_norm, native_swiglu, native_mlp_down, native_token_mixer) =
-        engine.trace_persistent_mlp_stage_after_layers(0, intermediate)?;
-    engine.rebuild_prefill_state(prefix_ids, true)?;
-    let _ = engine.decode_step_batch(trace_tokens, seqlen_offset)?;
-    let native_layer = engine
-        .state_for_batch(0)
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing native layer {trace_layer} after decode"))?;
-    let native_conv = native_layer
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing native conv state for layer {trace_layer}"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("native conv D2H layer {trace_layer}: {e}"))?;
-    let native_recurrent = native_layer
-        .recurrent_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing native recurrent state for layer {trace_layer}"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("native recurrent D2H layer {trace_layer}: {e}"))?;
-    let hidden_delta = replay_hidden
-        .map(|replay_hidden| {
-            validate::max_abs_delta(
-                &decode_bf16_le(&native_hidden),
-                &decode_bf16_le(replay_hidden),
-            )
-        })
-        .unwrap_or(0.0);
-    let replay_comp_trace = if let Some(replay_hidden) = replay_hidden {
-        engine.rebuild_prefill_state(prefix_ids, true)?;
-        engine.set_hidden_from_bytes(replay_hidden)?;
-        Some(engine.component_trace_linear_layer_from_current_hidden(trace_layer)?)
-    } else {
-        None
-    };
-    let comp_vs_replay_conv = validate::max_abs_delta(
-        &decode_bf16_le(&native_comp_conv),
-        &decode_bf16_le(&replay_conv),
-    );
-    let comp_vs_replay_recurrent = validate::max_abs_delta(
-        &decode_f32_le(&native_comp_recurrent),
-        &decode_f32_le(&replay_recurrent),
-    );
-    let comp_vs_replay_hidden = validate::max_abs_delta(
-        &decode_bf16_le(&native_comp_hidden),
-        &decode_bf16_le(replay_hidden_out),
-    );
-    let native_vs_comp_conv = validate::max_abs_delta(
-        &decode_bf16_le(&native_conv),
-        &decode_bf16_le(&native_comp_conv),
-    );
-    let native_vs_comp_recurrent = validate::max_abs_delta(
-        &decode_f32_le(&native_recurrent),
-        &decode_f32_le(&native_comp_recurrent),
-    );
-    let native_vs_comp_proj_residual = validate::max_abs_delta(
-        &decode_bf16_le(&native_hidden_out),
-        &bf16_residual_sum(&native_hidden, &native_comp_trace.proj_out),
-    );
-    let native_vs_comp_qkv_proj = validate::max_abs_delta(
-        &decode_f32_le(&native_qkv_proj),
-        &decode_bf16_le(&native_comp_trace.qkv),
-    );
-    let native_vs_comp_z_proj = validate::max_abs_delta(
-        &decode_f32_le(&native_z_proj),
-        &decode_bf16_le(&native_comp_trace.z),
-    );
-    let native_vs_comp_b_proj = validate::max_abs_delta(
-        &decode_f32_le(&native_b_proj),
-        &decode_bf16_le(&native_comp_trace.b),
-    );
-    let native_vs_comp_a_proj = validate::max_abs_delta(
-        &decode_f32_le(&native_a_proj),
-        &decode_bf16_le(&native_comp_trace.a),
-    );
-    let native_vs_comp_post_norm = validate::max_abs_delta(
-        &decode_f32_le(&native_post_norm),
-        &decode_bf16_le(&native_comp_layer.post_attn_norm),
-    );
-    let native_vs_comp_gated = validate::max_abs_delta(
-        &decode_f32_le(&native_gated),
-        &decode_bf16_le(&native_comp_trace.gated),
-    );
-    let native_vs_comp_swiglu = validate::max_abs_delta(
-        &decode_f32_le(&native_swiglu),
-        &decode_bf16_le(&native_comp_layer.mlp_swiglu),
-    );
-    let native_vs_comp_token_mixer = validate::max_abs_delta(
-        &decode_f32_le(&native_token_mixer),
-        &decode_bf16_le(&native_comp_layer.attn_hidden),
-    );
-    let native_vs_comp_mlp_down = validate::max_abs_delta(
-        &decode_f32_le(&native_mlp_down),
-        &decode_bf16_le(&native_comp_layer.mlp_out),
-    );
-    let conv_state_len = cfg.linear_conv_kernel_dim - 1;
-    let expected_conv_tail = {
-        let start = decode_bf16_le(&pre_step_conv);
-        let qkv = decode_bf16_le(&native_comp_trace.qkv);
-        let mut expected = vec![0.0f32; qkv_dim * conv_state_len];
-        for c in 0..qkv_dim {
-            let base = c * conv_state_len;
-            for t in 0..conv_state_len.saturating_sub(1) {
-                expected[base + t] = start[base + t + 1];
-            }
-            expected[base + conv_state_len - 1] = qkv[c];
-        }
-        expected
-    };
-    let native_conv_vs_expected_tail =
-        validate::max_abs_delta(&decode_bf16_le(&native_conv), &expected_conv_tail);
-    let comp_conv_vs_expected_tail =
-        validate::max_abs_delta(&decode_bf16_le(&native_comp_conv), &expected_conv_tail);
-    let replay_conv_vs_expected_tail =
-        validate::max_abs_delta(&decode_bf16_le(&replay_conv), &expected_conv_tail);
-    let native_conv_tap_deltas = {
-        let native = decode_bf16_le(&native_conv);
-        let mut deltas = vec![0.0f32; conv_state_len];
-        for c in 0..qkv_dim {
-            let base = c * conv_state_len;
-            for t in 0..conv_state_len {
-                deltas[t] = deltas[t].max((native[base + t] - expected_conv_tail[base + t]).abs());
-            }
-        }
-        deltas
-            .iter()
-            .map(|v| format!("{v:.6}"))
-            .collect::<Vec<_>>()
-            .join(",")
-    };
-    let native_qkv_proj_f32 = decode_f32_le(&native_qkv_proj);
-    let native_z_proj_f32 = decode_f32_le(&native_z_proj);
-    let comp_qkv_proj_f32 = decode_bf16_le(&native_comp_trace.qkv);
-    let comp_z_proj_f32 = decode_bf16_le(&native_comp_trace.z);
-    let (max_append_channel, max_append_mismatch) = {
-        let native = decode_bf16_le(&native_conv);
-        let start = decode_bf16_le(&pre_step_conv);
-        let qkv = decode_bf16_le(&native_comp_trace.qkv);
-        let conv_w = decode_bf16_le(
-            &engine
-                .weights()
-                .layers
-                .get(trace_layer)
-                .ok_or_else(|| anyhow::anyhow!("missing weights for layer {trace_layer}"))?
-                .linear
-                .as_ref()
-                .ok_or_else(|| anyhow::anyhow!("layer {trace_layer} is not linear"))?
-                .conv1d_w
-                .to_host_bytes()
-                .map_err(|e| anyhow::anyhow!("conv1d_w D2H layer {trace_layer}: {e}"))?,
-        );
-        let mut best = (0usize, 0.0f32);
-        for c in 0..qkv_dim {
-            let idx = c * conv_state_len + (conv_state_len - 1);
-            let delta = (native[idx] - expected_conv_tail[idx]).abs();
-            if delta > best.1 {
-                best = (c, delta);
-            }
-        }
-        let channel = best.0;
-        let idx = channel * conv_state_len + (conv_state_len - 1);
-        let weight_base = channel * cfg.linear_conv_kernel_dim;
-        let state_base = channel * conv_state_len;
-        let mut conv_acc = 0.0f32;
-        for tap in 0..cfg.linear_conv_kernel_dim {
-            let x = if tap + 1 == cfg.linear_conv_kernel_dim {
-                qkv[channel]
-            } else {
-                start[state_base + tap]
-            };
-            conv_acc += x * conv_w[weight_base + tap];
-        }
-        let conv_out = bf16_round(conv_acc * sigmoid_fast(conv_acc));
-        let native_last = native[idx];
-        let mut nearest_qkv = (0usize, f32::INFINITY);
-        for (i, &v) in native_qkv_proj_f32.iter().enumerate() {
-            let delta = (v - native_last).abs();
-            if delta < nearest_qkv.1 {
-                nearest_qkv = (i, delta);
-            }
-        }
-        (
-            channel,
-            format!(
-                "channel={channel},native={:.6},expected={:.6},prev_last={:.6},qkv_comp={:.6},qkv_native={:.6},conv_out={:.6},nearest_qkv=(channel={},value={:.6},delta={:.6}),delta={:.6}",
-                native[idx],
-                expected_conv_tail[idx],
-                start[idx],
-                qkv[channel],
-                native_qkv_proj_f32[channel],
-                conv_out,
-                nearest_qkv.0,
-                native_qkv_proj_f32[nearest_qkv.0],
-                nearest_qkv.1,
-                best.1
-            )
-        )
-    };
-    engine.rebuild_prefill_state(prefix_ids, true)?;
-    let step_b_debug_raw = engine.trace_persistent_linear_step_b_after_layers(
-        trace_tokens,
-        seqlen_offset,
-        trace_layer + 1,
-        trace_layer,
-        max_append_channel,
-    )?;
-    let step_b_debug = {
-        let debug = decode_f32_le(&step_b_debug_raw);
-        let native = decode_bf16_le(&native_conv);
-        let partial = decode_bf16_le(&native_partial_conv);
-        let start = decode_bf16_le(&pre_step_conv);
-        let base = max_append_channel * conv_state_len;
-        let idx = base + (conv_state_len - 1);
-        let state_values = debug
-            .iter()
-            .take(conv_state_len)
-            .map(|v| format!("{v:.6}"))
-            .collect::<Vec<_>>()
-            .join(",");
-        let step_b_last = debug.get(conv_state_len - 1).copied().unwrap_or_default();
-        let step_b_vs_expected = (step_b_last - expected_conv_tail[idx]).abs();
-        let final_vs_step_b = (native[idx] - step_b_last).abs();
-        let partial_vs_step_b = (partial[idx] - step_b_last).abs();
-        let partial_vs_expected = (partial[idx] - expected_conv_tail[idx]).abs();
-        format!(
-            "channel={max_append_channel},state=[{state_values}],qkv={:.6},conv_out={:.6},shift0_expected={:.6},shift1_expected={:.6},append_expected={:.6},step_b_vs_expected={:.6},partial_last={:.6},partial_vs_step_b={:.6},partial_vs_expected={:.6},final_vs_step_b={:.6}",
-            debug.get(conv_state_len).copied().unwrap_or_default(),
-            debug.get(conv_state_len + 1).copied().unwrap_or_default(),
-            start.get(base + 1).copied().unwrap_or_default(),
-            start.get(base + 2).copied().unwrap_or_default(),
-            expected_conv_tail[idx],
-            step_b_vs_expected,
-            partial[idx],
-            partial_vs_step_b,
-            partial_vs_expected,
-            final_vs_step_b,
-        )
-    };
-    let first_later_clobber = {
-        let native = decode_bf16_le(&native_conv);
-        let partial = decode_bf16_le(&native_partial_conv);
-        let base = max_append_channel * conv_state_len;
-        let idx = base + (conv_state_len - 1);
-        let step_b_last = partial[idx];
-        if (native[idx] - step_b_last).abs() == 0.0
-            || trace_layer + 1 >= text_config.num_hidden_layers
-        {
-            "none".to_string()
-        } else {
-            let mut lo = trace_layer + 1;
-            let mut hi = text_config.num_hidden_layers;
-            let mut hi_last = native[idx];
-            while lo + 1 < hi {
-                let mid = lo + (hi - lo) / 2;
-                engine.rebuild_prefill_state(prefix_ids, true)?;
-                let _ = engine.decode_step_batch_trace_hidden_after_layers(
-                    trace_tokens,
-                    seqlen_offset,
-                    mid,
-                    0,
-                )?;
-                let mid_conv = engine
-                    .state_for_batch(0)
-                    .layers
-                    .get(trace_layer)
-                    .ok_or_else(|| anyhow::anyhow!("missing binary-search layer {trace_layer}"))?
-                    .conv_state
-                    .as_ref()
-                    .ok_or_else(|| {
-                        anyhow::anyhow!("missing binary-search conv state for layer {trace_layer}")
-                    })?
-                    .to_host_bytes()
-                    .map_err(|e| {
-                        anyhow::anyhow!("binary-search conv D2H layer {trace_layer}: {e}")
-                    })?;
-                let mid_vals = decode_bf16_le(&mid_conv);
-                let mid_last = mid_vals[idx];
-                if (mid_last - step_b_last).abs() > 0.0 {
-                    hi = mid;
-                    hi_last = mid_last;
-                } else {
-                    lo = mid;
-                }
-            }
-            format!(
-                "after_layers={hi},clobber_layer={},partial_last={:.6},clobbered_last={:.6},delta={:.6}",
-                hi - 1,
-                step_b_last,
-                hi_last,
-                (hi_last - step_b_last).abs()
-            )
-        }
-    };
-    let pointer_debug = {
-        let state0 = engine.state_for_batch(0);
-        let trace_layer_state = state0
-            .layers
-            .get(trace_layer)
-            .ok_or_else(|| anyhow::anyhow!("missing trace layer state {trace_layer}"))?;
-        let final_layer_idx = text_config.num_hidden_layers.saturating_sub(1);
-        let final_layer_state = state0
-            .layers
-            .get(final_layer_idx)
-            .ok_or_else(|| anyhow::anyhow!("missing final layer state {final_layer_idx}"))?;
-        let trace_conv = trace_layer_state
-            .conv_state
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        let trace_rec = trace_layer_state
-            .recurrent_state
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        let final_k = final_layer_state
-            .kv_cache_k
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        let final_v = final_layer_state
-            .kv_cache_v
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        let final_shadow_k = final_layer_state
-            .kv_shadow_k
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        let final_shadow_v = final_layer_state
-            .kv_shadow_v
-            .as_ref()
-            .map(|b| b.as_ptr() as usize)
-            .unwrap_or(0);
-        format!(
-            "trace_conv=0x{trace_conv:x},trace_rec=0x{trace_rec:x},final_k=0x{final_k:x},final_v=0x{final_v:x},final_shadow_k=0x{final_shadow_k:x},final_shadow_v=0x{final_shadow_v:x},workspace=0x{:x}",
-            engine.scratch_debug_ptr(),
-        )
-    };
-    let isolated_tail_windows = {
-        let starts = [4usize, 5, 6, 7, 8];
-        let mut samples = Vec::new();
-        for &start_layer in &starts {
-            if start_layer >= text_config.num_hidden_layers {
-                continue;
-            }
-            let window_layers = text_config.num_hidden_layers - start_layer;
-            engine.rebuild_prefill_state(prefix_ids, true)?;
-            let pre_hidden = engine.decode_step_batch_trace_hidden_after_layers(
-                trace_tokens,
-                seqlen_offset,
-                start_layer,
-                0,
-            )?;
-            let before_conv = engine
-                .state_for_batch(0)
-                .layers
-                .get(trace_layer)
-                .ok_or_else(|| anyhow::anyhow!("missing pre-window layer {trace_layer}"))?
-                .conv_state
-                .as_ref()
-                .ok_or_else(|| {
-                    anyhow::anyhow!("missing pre-window conv state for layer {trace_layer}")
-                })?
-                .to_host_bytes()
-                .map_err(|e| anyhow::anyhow!("pre-window conv D2H layer {trace_layer}: {e}"))?;
-            let _ = engine.debug_decode_window_from_hidden_bf16(
-                &pre_hidden,
-                seqlen_offset,
-                start_layer,
-                window_layers,
-                0,
-            )?;
-            let after_conv = engine
-                .state_for_batch(0)
-                .layers
-                .get(trace_layer)
-                .ok_or_else(|| anyhow::anyhow!("missing post-window layer {trace_layer}"))?
-                .conv_state
-                .as_ref()
-                .ok_or_else(|| {
-                    anyhow::anyhow!("missing post-window conv state for layer {trace_layer}")
-                })?
-                .to_host_bytes()
-                .map_err(|e| anyhow::anyhow!("post-window conv D2H layer {trace_layer}: {e}"))?;
-            let before_vals = decode_bf16_le(&before_conv);
-            let after_vals = decode_bf16_le(&after_conv);
-            let base = max_append_channel * conv_state_len;
-            let idx = base + (conv_state_len - 1);
-            samples.push(format!(
-                "{}:{}:{:.6}",
-                start_layer,
-                window_layers,
-                (after_vals[idx] - before_vals[idx]).abs()
-            ));
-        }
-        samples.join(",")
-    };
-    let append_mismatch_samples = {
-        let native = decode_bf16_le(&native_conv);
-        let mut mismatches = Vec::with_capacity(qkv_dim);
-        for c in 0..qkv_dim {
-            let idx = c * conv_state_len + (conv_state_len - 1);
-            mismatches.push((
-                c,
-                native[idx],
-                expected_conv_tail[idx],
-                (native[idx] - expected_conv_tail[idx]).abs(),
-            ));
-        }
-        mismatches.sort_by(|a, b| b.3.partial_cmp(&a.3).unwrap_or(std::cmp::Ordering::Equal));
-        mismatches
-            .into_iter()
-            .take(8)
-            .map(|(channel, native_last, expected_last, delta)| {
-                let mut nearest_qkv = (0usize, f32::INFINITY);
-                for (i, &v) in native_qkv_proj_f32.iter().enumerate() {
-                    let qkv_delta = (v - native_last).abs();
-                    if qkv_delta < nearest_qkv.1 {
-                        nearest_qkv = (i, qkv_delta);
-                    }
-                }
-                format!(
-                    "c{channel}->q{} native={:.6} expected={:.6} self_q={:.6} match_q={:.6} match_delta={:.6} delta={:.6}",
-                    nearest_qkv.0,
-                    native_last,
-                    expected_last,
-                    native_qkv_proj_f32[channel],
-                    native_qkv_proj_f32[nearest_qkv.0],
-                    nearest_qkv.1,
-                    delta
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" | ")
-    };
-    let comp_conv_tap_deltas = {
-        let comp = decode_bf16_le(&native_comp_conv);
-        let mut deltas = vec![0.0f32; conv_state_len];
-        for c in 0..qkv_dim {
-            let base = c * conv_state_len;
-            for t in 0..conv_state_len {
-                deltas[t] = deltas[t].max((comp[base + t] - expected_conv_tail[base + t]).abs());
-            }
-        }
-        deltas
-            .iter()
-            .map(|v| format!("{v:.6}"))
-            .collect::<Vec<_>>()
-            .join(",")
-    };
-    let native_vs_replay_post_norm = validate::max_abs_delta(
-        &decode_f32_le(&native_post_norm),
-        &decode_bf16_le(replay_post),
-    );
-    let native_vs_replay_gated = replay_comp_trace
-        .as_ref()
-        .map(|trace| {
-            validate::max_abs_delta(
-                &decode_f32_le(&native_gated),
-                &decode_bf16_le(&trace.0.gated),
-            )
-        })
-        .unwrap_or(0.0);
-    let native_vs_replay_swiglu = validate::max_abs_delta(
-        &decode_f32_le(&native_swiglu),
-        &decode_bf16_le(replay_swiglu),
-    );
-    let native_vs_replay_token_mixer = validate::max_abs_delta(
-        &decode_f32_le(&native_token_mixer),
-        &decode_bf16_le(replay_attn),
-    );
-    let native_vs_replay_mlp_down = validate::max_abs_delta(
-        &decode_f32_le(&native_mlp_down),
-        &decode_bf16_le(replay_mlp_out),
-    );
-    let native_vs_replay_qkv_proj = replay_comp_trace
-        .as_ref()
-        .map(|trace| {
-            validate::max_abs_delta(
-                &decode_f32_le(&native_qkv_proj),
-                &decode_bf16_le(&trace.0.qkv),
-            )
-        })
-        .unwrap_or(0.0);
-    let native_vs_replay_z_proj = replay_comp_trace
-        .as_ref()
-        .map(|trace| {
-            validate::max_abs_delta(&decode_f32_le(&native_z_proj), &decode_bf16_le(&trace.0.z))
-        })
-        .unwrap_or(0.0);
-    let native_vs_replay_b_proj = replay_comp_trace
-        .as_ref()
-        .map(|trace| {
-            validate::max_abs_delta(&decode_f32_le(&native_b_proj), &decode_bf16_le(&trace.0.b))
-        })
-        .unwrap_or(0.0);
-    let native_vs_replay_a_proj = replay_comp_trace
-        .as_ref()
-        .map(|trace| {
-            validate::max_abs_delta(&decode_f32_le(&native_a_proj), &decode_bf16_le(&trace.0.a))
-        })
-        .unwrap_or(0.0);
-    let comp_layer_vs_replay_hidden = validate::max_abs_delta(
-        &decode_bf16_le(&native_comp_layer.layer_hidden),
-        &decode_bf16_le(replay_hidden_out),
-    );
-    let native_vs_comp_layer_hidden = validate::max_abs_delta(
-        &decode_bf16_le(&native_hidden_out),
-        &decode_bf16_le(&native_comp_layer.layer_hidden),
-    );
-    let native_vs_replay_hidden = validate::max_abs_delta(
-        &decode_bf16_le(&native_hidden_out),
-        &decode_bf16_le(replay_hidden_out),
-    );
-    let sample_qkv_native = native_qkv_proj_f32
-        .iter()
-        .take(4)
-        .map(|v| format!("{v:.4}"))
-        .collect::<Vec<_>>()
-        .join(",");
-    let sample_qkv_comp = comp_qkv_proj_f32
-        .iter()
-        .take(4)
-        .map(|v| format!("{v:.4}"))
-        .collect::<Vec<_>>()
-        .join(",");
-    let sample_z_native = native_z_proj_f32
-        .iter()
-        .take(4)
-        .map(|v| format!("{v:.4}"))
-        .collect::<Vec<_>>()
-        .join(",");
-    let sample_z_comp = comp_z_proj_f32
-        .iter()
-        .take(4)
-        .map(|v| format!("{v:.4}"))
-        .collect::<Vec<_>>()
-        .join(",");
-
-    eprintln!(
-        "[trace-persistent-linear] layer={trace_layer} hidden_delta={hidden_delta:.6} comp_vs_replay_conv={comp_vs_replay_conv:.6} comp_vs_replay_recurrent={comp_vs_replay_recurrent:.6} comp_linear_hidden_vs_replay={comp_vs_replay_hidden:.6} native_vs_comp_qkv_proj={native_vs_comp_qkv_proj:.6} native_vs_replay_qkv_proj={native_vs_replay_qkv_proj:.6} native_vs_comp_z_proj={native_vs_comp_z_proj:.6} native_vs_replay_z_proj={native_vs_replay_z_proj:.6} native_vs_comp_b_proj={native_vs_comp_b_proj:.6} native_vs_replay_b_proj={native_vs_replay_b_proj:.6} native_vs_comp_a_proj={native_vs_comp_a_proj:.6} native_vs_replay_a_proj={native_vs_replay_a_proj:.6} native_vs_comp_conv={native_vs_comp_conv:.6} native_vs_comp_recurrent={native_vs_comp_recurrent:.6} native_conv_vs_expected_tail={native_conv_vs_expected_tail:.6} comp_conv_vs_expected_tail={comp_conv_vs_expected_tail:.6} replay_conv_vs_expected_tail={replay_conv_vs_expected_tail:.6} native_conv_tap_deltas=[{native_conv_tap_deltas}] comp_conv_tap_deltas=[{comp_conv_tap_deltas}] max_append_mismatch=({max_append_mismatch}) step_b_debug=({step_b_debug}) first_later_clobber=({first_later_clobber}) pointer_debug=({pointer_debug}) isolated_tail_windows=[{isolated_tail_windows}] append_mismatch_samples=[{append_mismatch_samples}] native_vs_comp_token_mixer={native_vs_comp_token_mixer:.6} native_vs_replay_token_mixer={native_vs_replay_token_mixer:.6} native_vs_comp_post_norm={native_vs_comp_post_norm:.6} native_vs_replay_post_norm={native_vs_replay_post_norm:.6} native_vs_comp_gated={native_vs_comp_gated:.6} native_vs_replay_gated={native_vs_replay_gated:.6} native_vs_comp_swiglu={native_vs_comp_swiglu:.6} native_vs_replay_swiglu={native_vs_replay_swiglu:.6} native_vs_comp_mlp_down={native_vs_comp_mlp_down:.6} native_vs_replay_mlp_down={native_vs_replay_mlp_down:.6} native_vs_comp_proj_residual={native_vs_comp_proj_residual:.6} comp_layer_hidden_vs_replay={comp_layer_vs_replay_hidden:.6} native_vs_comp_layer_hidden={native_vs_comp_layer_hidden:.6} native_vs_replay_hidden={native_vs_replay_hidden:.6} sample_qkv_native=[{sample_qkv_native}] sample_qkv_comp=[{sample_qkv_comp}] sample_z_native=[{sample_z_native}] sample_z_comp=[{sample_z_comp}]"
-    );
     Ok(())
 }
